add comment count changed event

+++ services/fly-comment-service/src/main/java/com/fyp/fly/services/comment/service/DefaultCommentService.java
@@ -1,11 +1,16 @@
 package com.fyp.fly.services.comment.service;
 
+import com.fyp.fly.common.enums.CountBizType;
+import com.fyp.fly.common.event.CountEvent;
+import com.fyp.fly.common.event.FlyEvent;
 import com.fyp.fly.common.result.api.JsonResult;
 import com.fyp.fly.common.result.api.ResultUtils;
+import com.fyp.fly.common.utils.JSONUtils;
 import com.fyp.fly.services.comment.domain.Comment;
 import com.fyp.fly.services.comment.domain.CommentDto;
 import com.fyp.fly.services.comment.repository.mapper.CommentMapper;
 import org.apache.commons.lang.StringUtils;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +20,9 @@ public class DefaultCommentService implements CommentService {
     @Autowired
     private CommentMapper commentMapper;
 
+    @Autowired
+    private RabbitTemplate rabbitTemplate;
+
     @Override
     public JsonResult add(CommentDto comment) {
         if (comment == null) {
@@ -23,6 +31,7 @@ public class DefaultCommentService implements CommentService {
         if (comment.isValid()) {
             Comment commentModel = comment.transfer();
             commentMapper.add(commentModel);
+            sendCommentCountChangedEvent(comment.getArtId(),true);
             return ResultUtils.success(commentModel.getId());
         } else {
             return ResultUtils.failed(comment.getErrMsg());
@@ -34,7 +43,12 @@ public class DefaultCommentService implements CommentService {
         if (id == null) {
             throw new IllegalArgumentException("id");
         }
+        Long artId = commentMapper.getArtIdById(id);
+        if (artId == null){
+            return ResultUtils.failed("帖子不存在");
+        }
         commentMapper.delete(id);
+        sendCommentCountChangedEvent(artId,false);
         return ResultUtils.success();
     }
 
@@ -54,5 +68,15 @@ public class DefaultCommentService implements CommentService {
         }
         commentMapper.updateContent(id,uid,content);
         return ResultUtils.success();
+    }
+
+    private void sendCommentCountChangedEvent(Long id,boolean increment) {
+
+        CountEvent event = new CountEvent();
+        event.setBizId(id);
+        event.setBizType(CountBizType.ARTICLE_COMMENT.getCode());
+        event.setIncrement(increment);
+
+        rabbitTemplate.convertAndSend(FlyEvent.SERVICE_COMMON_EXCHANGE, FlyEvent.SERVICE_ARTICLE_COUNT_EVENT, JSONUtils.toJSONString(event));
     }
 }
